lexer done
