+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: caigner <caigner@student.42.fr>            +#+  +:+       +#+         #
+#    By: miheider <miheider@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/03/10 20:04:47 by miheider          #+#    #+#              #
-#    Updated: 2024/05/12 17:24:29 by caigner          ###   ########.fr        #
+#    Updated: 2024/05/12 22:01:40 by miheider         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -37,6 +37,15 @@ SOURCES	=	src/builtins/ft_echo.c	\
 			src/parsing_3.c \
 			src/open_io_1.c \
 			src/open_io_2.c \
+			src/lexer_1.c \
+			src/lexer_2.c \
+			src/lexer_3.c \
+			src/lexer_4.c \
+			src/lexer_5.c \
+			src/lexer_6.c \
+			src/lexer_7.c \
+			src/lexer_8.c \
+			src/lexer_9.c \
 			src/get_cmd_path_1.c \
 			src/get_cmd_path_2.c \
 			src/exec.c \
